+++ src/apis/comment.js
@@ -1,10 +1,10 @@
-import { POSTURL } from '../utils/constant/post';
+import { POST_URL } from '../utils/constants/post';
 import { authRequest } from './core';
 
 // 댓글 생성하기
 export const createComment = async (postId, comment) => {
   try {
-    return await authRequest.post(POSTURL.CREATE_COMMENTS, {
+    return await authRequest.post(POST_URL.CREATE_COMMENTS, {
       comment,
       postId,
     });
