+++ src/apis/search.js
@@ -1,8 +1,8 @@
-import { POSTURL } from '../utils/constant/post';
+import { POST_URL } from '../utils/constants/post';
 import { baseRequest } from './core';
 export const searchAll = async (keyword) => {
   try {
-    const data = await baseRequest.get(POSTURL.SEARCH_POST + keyword);
+    const data = await baseRequest.get(POST_URL.SEARCH_POST + keyword);
     return data;
   } catch (error) {
     console.error(error);
