+++ src/components/MyHome/UserInfo/index.jsx
@@ -1,14 +1,12 @@
 import { useEffect } from 'react';
 import { UserInfoContainer, UserInfoWrapper } from './style';
-import { useGetMyhomeUserInfo } from '../../../hooks/useGetMyhomeUserInfo';
-import { getStorage } from '../../../utils/storage';
-import { USERIMAGE } from '../../../utils/constant/auth';
+import { useGetMyHomeUserInfo } from '../../../hooks/useGetMyHomeUserInfo';
 import Avatar from '../../common/Avatar';
 import styled from '@emotion/styled';
 
 const UserInfo = () => {
-  const { profile, getUserData } = useGetMyhomeUserInfo();
-  const userImage = getStorage(USERIMAGE);
+  const { profile, getUserData } = useGetMyHomeUserInfo();
+
   useEffect(() => {
     getUserData();
   }, []);
