+++ src/components/MyHome/index.jsx
@@ -1,7 +1,7 @@
 import { Fieldset, FormButton, FormLogin, LoginContainer, LoginWrapper } from '../Signin/style';
 import { FormSettingText, Input, PasswordBlock, PasswordText } from './style';
 import { useNewPassWordForm } from '../../hooks/useNewPassWordForm';
-import { USER as AUTH } from '../../utils/constant/auth';
+import { USER as AUTH } from '../../utils/constants/auth';
 
 const UserSettingPassword = () => {
   const [handleChange, handleSubmit, values, passwordError] = useNewPassWordForm();
