+++ src/hooks/useClickAway.js
@@ -1,12 +1,10 @@
-import { useRef, useEffect } from 'react';
+import { useEffect, useRef } from 'react';
 import { EVENT } from '../utils/constant/hooks';
-// 추후에 constant로 옮겨도 될듯
 
 const useClickAway = (handler) => {
   const ref = useRef(null);
   const savedHandler = useRef(handler);
 
-  // 핸들러 함수가 변하더라도 리렌더링 안시키기 위함
   useEffect(() => {
     savedHandler.current = handler;
   }, [handler]);
@@ -16,7 +14,6 @@ const useClickAway = (handler) => {
     if (!element) return;
 
     const handleEvent = (e) => {
-      // element를 바로 사용하면, dom이 바뀐게 추적이 안됨. (image의 src같은 부분)
       !ref.current.contains(e.target) && savedHandler.current(e);
     };
