+++ src/hooks/useLike.js
@@ -1,13 +1,13 @@
 import { useEffect, useState } from 'react';
 import { useRecoilState } from 'recoil';
+import { createAlarm } from '../apis/alarm';
+import { createLike, deleteLike } from '../apis/like';
 import { postStateFamily } from '../recoil/postStates';
 import { getStorage } from '../utils/storage';
-import { createLike, deleteLike } from '../apis/like';
-import { createAlarm } from '../apis/alarm';
 
 const useLike = (postId, author, likes) => {
-  const userId = getStorage('id'); //내 아이디
-  const [{ post }, setPost] = useRecoilState(postStateFamily(postId)); //동행 클릭한 post의 정보
+  const userId = getStorage('id');
+  const [{ post }, setPost] = useRecoilState(postStateFamily(postId));
   const [currentLike, setCurrentLike] = useState(likes.find(({ user }) => user === userId));
   const [isLike, setIsLike] = useState(false);
 
@@ -18,7 +18,6 @@ const useLike = (postId, author, likes) => {
   const onClickLike = async (e) => {
     e.stopPropagation();
 
-    //동행 취소
     if (isLike) {
       const data = await deleteLike(currentLike._id);
       setIsLike(false);
@@ -28,9 +27,7 @@ const useLike = (postId, author, likes) => {
 
         setPost({ key: postId, post: { ...post, likes: newPostLike } });
       }
-    }
-    // 동행 신청
-    else {
+    } else {
       const data = await createLike(postId);
       setIsLike(true);
 
@@ -41,7 +38,6 @@ const useLike = (postId, author, likes) => {
         setPost({ key: postId, post: { ...post, likes: newPostLike } });
       }
 
-      //알림 생성
       await createAlarm('LIKE', data._id, author._id, data.post);
 
       setCurrentLike(data);
