+++ src/hooks/useNewPassWordForm.js
@@ -1,5 +1,5 @@
 import { useState } from 'react';
-import { putPaswwordChange } from '../apis/auth';
+import { putPasswordChange } from '../apis/auth';
 import { ERROR_MESSAGE_SIGNIN, USER } from '../utils/constant/auth';
 
 export const useNewPassWordForm = () => {
@@ -21,7 +21,7 @@ export const useNewPassWordForm = () => {
   const handleSubmit = async (e) => {
     e.preventDefault();
 
-    const res = await putPaswwordChange(values.password);
+    const res = await putPasswordChange(values.password);
 
     if (res.status === 200) {
       swal(USER.CHANGE_PASSWORD_SUCCESS, '', ERROR_MESSAGE_SIGNIN.PASSWORD_SUCCESS);
