+++ src/hooks/usePoster.js
@@ -1,14 +1,14 @@
-import { useRecoilState } from 'recoil';
-import { myhomeModalState, uploadImageState, userLoginDateState } from '../recoil/uploadImageState';
-import { getMyPostDetail, removePost } from '../apis/post';
-import { ERROR_MESSAGE_SIGNIN, USER } from '../utils/constant/auth';
-import { getUser } from '../apis/auth';
 import { useEffect, useState } from 'react';
+import { useRecoilState } from 'recoil';
+import { getUser } from '../apis/auth';
+import { getMyPostDetail, removePost } from '../apis/post';
+import { myHomeModalState, uploadImageState, userLoginDateState } from '../recoil/uploadImageState';
+import { ERROR_MESSAGE_SIGNIN, USER } from '../utils/constant/auth';
 
 export const usePoster = () => {
   const [getLoginData, setLoginData] = useState({});
   const [postId, setPostId] = useState('');
-  const [visible, setVisible] = useRecoilState(myhomeModalState);
+  const [visible, setVisible] = useRecoilState(myHomeModalState);
 
   const [postDetail, setPostDetail] = useRecoilState(userLoginDateState);
 
@@ -18,10 +18,10 @@ export const usePoster = () => {
   const [loading, setLoading] = useState(false);
 
   const handlePoster = async (id) => {
-    const getpostDetail = await getMyPostDetail(id);
+    const getPostDetail = await getMyPostDetail(id);
     setVisible(true);
     setPostId(id);
-    setPostDetail(getpostDetail);
+    setPostDetail(getPostDetail);
   };
 
   const handlerModalClose = () => {
