--- src/utils/constant/hooks.js
@@ -1,1 +1,0 @@
-export const EVENT = ['mousedown', 'touchstart'];
