--- src/utils/constant/post.js
@@ -1,29 +1,0 @@
-export const ADD_COMMENT_PLACEHOLDER = '댓글달기...';
-
-export const ERROR_MESSAGE_POST = {
-  ERROR_CHANNELS: '채널 불러오기 실패',
-  ERROR_POSTS: '포스트 불러오기 실패',
-  ERROR_CREATE_POST: '포스트 생성 실패',
-  ERROR_POSTDETAIL: '포스트 상세 불러오기 실패',
-};
-export const POSTURL = {
-  CHANNELS: '/channels',
-  POSTS_IN_CHANNEL: '/posts/channel/',
-  CREATE_POST: 'posts/create',
-  ALL_POST: '/posts',
-  POST_DETAIL: '/posts/',
-  CREATE_COMMENTS: '/comments/create',
-  CREATE_LIKES: '/likes/create',
-  DELETE_LIKES: '/likes/delete',
-  SEARCH_POST: '/search/all/',
-};
-
-export const ERROR_MESSAGE_POSTMODAL = {
-  country: '나라를 선택해주세요.',
-  startDate: '동행 시작 날짜를 선택해주세요.',
-  endDate: '동행 끝 날짜를 선택해주세요.',
-  personnel: '인원을 입력해주세요.',
-  gender: '원하는 성별을 선택해주세요.',
-  title: '제목을 입력해주세요.',
-  content: '내용을 입력해주세요.',
-};
