--- src/utils/constant/user.js
@@ -1,5 +1,0 @@
-export const USER_URL = {
-  GET_ALL_USER: '/users/get-users',
-  GET_SEARCHED_USER: '/search/users/',
-  GET_CLICKED_USER: '/users/',
-};
