+++ src/utils/constants/alarm.js
@@ -1,0 +1,9 @@
+export const ALARM = {
+  COMMENT: '댓글을',
+  ACCOMPANY: '동행신청을',
+};
+
+export const URL = {
+  GET_ALARM: 'notifications',
+  CREATE_ALARM: '/notifications/create',
+};
