+++ src/utils/constants/hooks.js
@@ -1,0 +1,1 @@
+export const EVENT = ['mousedown', 'touchstart'];
