--- src/utils/myhome/constant.js
@@ -1,18 +1,0 @@
-export const FILE = {
-  JPEG: 'image/jpeg',
-  PNG: 'image/png',
-  GIF: 'image/gif',
-  SVG: 'image/svg+xml',
-};
-
-export const FORMATDATA = {
-  POST_ID: 'postId',
-  IMAGE: 'image',
-  TITLE: 'title',
-  CHANNEL_ID: 'channelId',
-};
-
-export const ERROR_MESSAGE = {
-  UPLOAD: '사진을 업로드 해주세요',
-  UPLOAD_IMAGE: '이미지 파일만 업로드 가능합니다',
-};
